#121 style: Rename user slice to auth slice redux에서 쓰이는 user를 auth로 변경, 기타 css 수정
