+++ src/components/Common/CalendarContainer.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect, useRef } from "react";
 import { useSelector, useDispatch } from "react-redux";
 
+import { setEdit } from "@/features/auth/auth-slice";
 import {
 	currentMonthFn,
 	currentYearFn,
 	setId,
 } from "@/features/schedule/schedule-slice";
 import { openModal } from "@/features/ui/ui-slice";
-import { setEdit } from "@/features/user/user-slice";
 import { getRandomColor } from "@/utils/color";
 
 import CustomCalendar from "./CustomCalendar/CustomCalendar";
