+++ src/components/PersonalTodoList/PersonalTodoList/PersonalTodoList.jsx
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 
 import ScheduleModal from "@/components/Common/ScheduleModal/ScheduleModal.jsx";
+import { setEdit } from "@/features/auth/auth-slice.js";
 import { getSchedule } from "@/features/schedule/schedule-service.js";
 import { openModal } from "@/features/ui/ui-slice";
-import { setEdit } from "@/features/user/user-slice.js";
 
 import {
 	TodoContainer,
