+++ src/components/Setting/PasswordUpdate.jsx
@@ -4,7 +4,7 @@ import { useNavigate } from "react-router-dom";
 
 import { Typography, Button, TextField, Box, Grid, Alert } from "@mui/material";
 
-import { updateUserPassword } from "@/features/user/user-service";
+import { updateUserPassword } from "@/features/auth/auth-service";
 
 import { UserInfoContainer } from "./SettingPageDetail.styles";
