+++ src/components/SharePage/ShareTodoList/ShareTodoList.jsx
@@ -3,10 +3,10 @@ import { useDispatch, useSelector } from "react-redux";
 
 import ScheduleModal from "@/components/Common/ScheduleModal/ScheduleModal.jsx";
 import PersonalTodoItem from "@/components/PersonalTodoList/PersonalTodoItem/PersonalTodoItem.jsx";
+import { setEdit } from "@/features/auth/auth-slice.js";
 import { createInviteLink } from "@/features/group/group-invite-service.js";
 import { getSchedule } from "@/features/schedule/schedule-service.js";
 import { openModal } from "@/features/ui/ui-slice.js";
-import { setEdit } from "@/features/user/user-slice.js";
 
 import {
 	TodoContainer,
