+++ src/pages/FeedPage.jsx
@@ -2,8 +2,5 @@ import React from "react";
 
 import FeedMain from "@/components/FeedPage/FeedMain/FeedMain";
 
-const FeedPage = () => {
-	return <FeedMain />;
-};
-
+const FeedPage = () => <FeedMain />;
 export default FeedPage;
